Write separate files 'dorsal_configure' and 'dorsal_build'.
I find it helpful to be able to re-run and experiment with the
config and build steps directly when getting new (or changed)
packages to build.

Also takes out the logging of the build step, because many of
the configure/build scripts uses fancy tty features (colours,
redrawn lines) which do not work with pipes.